norm: separate parse_list*() functions
