+++ Makefile
@@ -24,6 +24,7 @@ SRCS		=	$(SRCS_DIR)main.c	\
 				$(SRCS_DIR)tokenize_print_token.c	\
 				$(SRCS_DIR)tokenize_utils.c	\
 				$(SRCS_DIR)parse.c	\
+				$(SRCS_DIR)parse_list.c	\
 				$(SRCS_DIR)parse_redirection.c	\
 				$(SRCS_DIR)parse_print_tree.c	\
 				$(SRCS_DIR)traverse.c	\
