+++ src/parse.c
@@ -121,51 +121,6 @@ t_tree	*parse_pipeline(t_list **tokens)
 	return (tree);
 }
 
-t_tree	*parse_list_end(t_tree *tree)
-{
-	tree->category = TR_LIST_END;
-	tree->right = NULL;
-	return (tree);
-}
-
-t_tree	*parse_list_and(t_list **tokens, t_tree *tree)
-{
-	tree->category = TR_LIST_AND;
-	*tokens = (*tokens)->next;
-	tree->right = parse_list(tokens);
-	return (tree);
-}
-
-t_tree	*parse_list_or(t_list **tokens, t_tree *tree)
-{
-	tree->category = TR_LIST_OR;
-	*tokens = (*tokens)->next;
-	tree->right = parse_list(tokens);
-	return (tree);
-}
-
-t_tree	*parse_list(t_list **tokens)
-{
-	t_tree	*tree;
-	t_token	*token;
-
-	tree = null_guard(malloc(sizeof(t_tree)),
-			PROGRAM_NAME, "parse_list().");
-	tree->left = parse_pipeline(tokens);
-	tree->right = NULL;
-	if (*tokens == NULL)
-		return (parse_list_end(tree));
-	token = (*tokens)->content;
-	if (token->category == T_AND)
-		return (parse_list_and(tokens, tree));
-	else if (token->category == T_OR)
-		return (parse_list_or(tokens, tree));
-	else if (token->category == T_R_PAREN)
-		return (parse_list_end(tree));
-	panic("parse_list()");
-	return (NULL);
-}
-
 t_tree	*parse(t_list *tokens)
 {
 	t_tree	*tree;
